v0.4.131: shared suite-editor helpers — collapse assignment/draft handler duplication
Every :assignmentID-scoped editor handler had a draft-scoped sibling
at /instructor/new/draft/... that duplicated auth, setup resolution,
DTO translation, and JSON response building.  That duplication was
the structural reason the create page is multiple versions behind
the edit page on Sections, Notebook Checks, and Support Files.  New
SuiteEditHelpers.swift exposes requireInstructor, loadAssignmentAndSetup,
loadDraftSetup, applySuiteEdit, applyPatternFamiliesEdit, and
applyNotebookChecksEdit; the five route files collapse to thin
handlers that resolve their target and call into the shared core.

Side effect: drafts now correctly persist sectionID on suite items
through PUT /draft/suite (was silently dropped pre-fix because the
draft handler used the pre-v0.4.96 payload shape).
