+++ CHANGELOG.md
@@ -6,6 +6,44 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+## [0.4.131] - 2026-04-29
+
+### Refactor
+
+- **Shared core for the suite / families / checks / suite-sections
+  endpoints.**  Pre-fix, every `:assignmentID`-scoped editor handler had
+  a draft-scoped sibling at `/instructor/new/draft/...` that duplicated
+  the auth check, setup resolution, body decoding, DTO translation, and
+  JSON response building.  That duplication was the structural reason
+  the create page is multiple versions behind the edit page on
+  Sections, Notebook Checks, and Support Files: each new feature on the
+  edit side meant writing (and forgetting) a parallel draft handler.
+  The new `Sources/APIServer/Routes/Web/SuiteEditHelpers.swift` exposes
+  `requireInstructor`, `loadAssignmentAndSetup`, `loadDraftSetup`,
+  `applySuiteEdit`, `applyPatternFamiliesEdit`, `applyNotebookChecksEdit`,
+  and `jsonResponse` — `AssignmentRoutes+Suite.swift`,
+  `AssignmentRoutes+Families.swift`, `AssignmentRoutes+Checks.swift`,
+  `AssignmentRoutes+SuiteSections.swift`, and `AssignmentRoutes+Draft.swift`
+  all collapse to thin handlers that resolve their target and call into
+  the shared core.  Net: ~150 fewer lines, and adding a draft-scoped
+  Checks / Sections / SectionVariables endpoint is now a few lines of
+  routing rather than a duplicate handler.
+
+### Fixed
+
+- **Drafts now persist `sectionID` on suite items through `PUT /draft/suite`**
+  and preserve pattern-family `variables` on row-level edits.  Both
+  fields landed on the assignment-scoped path
+  (`AuthoredSuiteItem.sectionID` in v0.4.96, `PatternFamily.variables`
+  in v0.4.94) but the draft-side handler still used the pre-v0.4.96
+  payload shape — a section assignment made via the suite editor on
+  the create page would silently drop on save.  Routing draft saves
+  through `applySuiteEdit` (the same shared core the assignment-scoped
+  handler uses) closes the gap as a side effect of the refactor.  No
+  draft data has been lost; pre-fix the field simply wasn't accepted on
+  save (the create page hasn't shipped a Sections UI yet, so this only
+  manifested for clients sending the field directly).
+
 ## [0.4.130] - 2026-04-29
 
 ### Fixed
