+++ Sources/APIServer/Routes/Web/AssignmentRoutes+Checks.swift
@@ -21,14 +21,8 @@ extension AssignmentRoutes {
 
     @Sendable
     func getNotebookChecks(req: Request) async throws -> Response {
-        let user = try req.auth.require(APIUser.self)
-        guard user.isInstructor else { throw Abort(.forbidden) }
-
-        let idStr = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(idStr, on: req.db),
-            let setup      = try await APITestSetup.find(assignment.testSetupID, on: req.db)
-        else { throw Abort(.notFound) }
+        try requireInstructor(req)
+        let (_, setup) = try await loadAssignmentAndSetup(req)
 
         let checks: [NotebookCheck] = {
             guard let data = setup.manifest.data(using: .utf8),
@@ -37,12 +31,7 @@ extension AssignmentRoutes {
             return props.notebookChecks
         }()
 
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        let data = try encoder.encode(checks)
-        return Response(status: .ok,
-                        headers: ["Content-Type": "application/json"],
-                        body: .init(data: data))
+        return try jsonResponse(checks)
     }
 
     // MARK: - PUT /instructor/:assignmentID/checks
@@ -57,14 +46,8 @@ extension AssignmentRoutes {
 
     @Sendable
     func putNotebookChecks(req: Request) async throws -> Response {
-        let user = try req.auth.require(APIUser.self)
-        guard user.isInstructor else { throw Abort(.forbidden) }
-
-        let idStr = try assignmentPublicIDParameter(from: req)
-        guard
-            let assignment = try await assignmentByPublicID(idStr, on: req.db),
-            let setup      = try await APITestSetup.find(assignment.testSetupID, on: req.db)
-        else { throw Abort(.notFound) }
+        try requireInstructor(req)
+        let (assignment, setup) = try await loadAssignmentAndSetup(req)
 
         let checks: [NotebookCheck]
         do {
@@ -74,30 +57,10 @@ extension AssignmentRoutes {
                 reason: "Invalid notebook check list: \(error.localizedDescription)")
         }
 
-        // Carry forward the current families list — this endpoint only
-        // edits checks.  The shared apply path validates and rewrites the
-        // manifest with both lists in a single zip-mutation pass.
-        let currentFamilies: [PatternFamily] = {
-            guard let data = setup.manifest.data(using: .utf8),
-                  let props = try? JSONDecoder().decode(TestProperties.self, from: data)
-            else { return [] }
-            return props.patternFamilies
-        }()
-
-        _ = try await applyPatternFamilies(
-            to: setup,
-            nextFamilies: currentFamilies,
-            nextChecks: checks,
-            on: req.db
-        )
+        try await applyNotebookChecksEdit(setup: setup, checks: checks, on: req.db)
 
         await scheduleValidationAfterSuiteEdit(req: req, assignment: assignment)
 
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        let data = try encoder.encode(checks)
-        return Response(status: .ok,
-                        headers: ["Content-Type": "application/json"],
-                        body: .init(data: data))
+        return try jsonResponse(checks)
     }
 }
